add navbar items and animate when hovering
Each item in the <ul> holds a navbar item. Use flexbox to center and
style things. Hide the text when we aren't hover, then on hover, make it
display again using `display: block`.

The <ul> (`.nav-list`) is using flexbox and has the direction set to
column so items appear top to bottom. The last item uses the
`:last-child` psuedo-selector and uses the `auto` value on `margin-top`
to push the item all the way to the bottom by making its top margin as
large as possible.